requirements.txt
absl-py=0.6.1
astor=0.7.1
blas=1.0
bzip2=1.0.6
c-ares=1.15.0
ca-certificates=2018.10.15
certifi=2018.10.15
cffi=1.11.5
cudatoolkit=9.0
cudnn=7.1.2
freetype=2.9.1
gast=0.2.0
grpcio=1.14.1
intel-openmp=2019.0
jpeg=9b
libedit=3.1.20170329
libffi=3.2.1
libgcc-ng=8.2.0
libgfortran-ng=7.3.0
libpng=1.6.35
libprotobuf=3.6.1
libstdcxx-ng=8.2.0
libtiff=4.0.9
markdown=2.6.11
mkl=2019.0
mkl_fft=1.0.6
mkl_random=1.0.2
nccl=1.3.5
ncurses=6.1
ninja=1.8.2
numpy=1.15.4
numpy-base=1.15.4
olefile=0.46
openssl=1.0.2p
pillow=5.3.0
pip=18.1
protobuf=3.6.1
pycparser=2.19
python=3.6.6
pytorch=0.4.1
readline=7.0
setuptools=40.6.2
six=1.11.0
sqlite=3.25.2
tensorboard=1.10.0
tensorflow=1.10.0
termcolor=1.1.0
tk=8.6.8
torchvision=0.2.1
werkzeug=0.14.1
wheel=0.32.2
xz=5.2.4
zlib=1.2.11
